Add Footer.js and Main.js
