+++ src/Footer.js
@@ -1,0 +1,9 @@
+import { Component } from "react";
+
+class Footer extends Component {
+  render() {
+    return <h1>I am a footer</h1>;
+  }
+}
+
+export default Footer;
